arrange files and edit directory according to user
