+++ Admin/addmanagerclinic.php
@@ -26,7 +26,7 @@ function getManagerRegion(val) {
 
 </script>
 </head>
-<body background= "clinicview.jpg">
+<body background= "../images/clinicview.jpg">
 <ul>
 <li class="dropdown"><font color="yellow" size="10">ADMIN MODE</font></li>
 <br>
@@ -75,7 +75,7 @@ function getManagerRegion(val) {
 		<select name="city" id="city-list" class="demoInputBox"  onChange="getState(this.value);getManagerRegion(this.value);">
 		<option value="">Select City</option>
 		<?php
-		include 'dbconfig.php';
+		include '../dbconfig.php';
 		$sql1="SELECT distinct City FROM clinic";
          $results=$conn->query($sql1); 
 		while($rs=$results->fetch_assoc()) { 
@@ -108,7 +108,7 @@ if(isset($_POST['logout'])){
 	}
 if(isset($_POST['Submit']))
 {
-		include 'dbconfig.php';
+		include '../dbconfig.php';
 		$cid=$_POST['clinic'];
 		$mid=$_POST['manager'];
