--- Admin/dbconfig.php
@@ -1,5 +1,0 @@
-				<?php
-
-				$conn = new mysqli('localhost', 'root', '', 'wt_database') 
-					or die ('Cannot connect to db');
-				?>
