+++ Admin/showclinic.php
@@ -25,7 +25,7 @@ tr,td{
 </style>
 
 </head>
-<body background= "clinicview.jpg">
+<body background= "../images/clinicview.jpg">
 <ul>
 <li class="dropdown"><font color="yellow" size="10">ADMIN MODE</font></li>
 <br>
@@ -70,19 +70,15 @@ tr,td{
 </h2>
 <center><h1>SHOW CLINIC</h1><hr>
 <?php
+include '../dbconfig.php';
 session_start();
 if(isset($_POST['logout'])){
 		session_unset();
 		session_destroy();
-		header( "Refresh:1; url=alogin.php"); 
+		header( "Refresh:1; url=../cover.php"); 
 	}
-$con = mysqli_connect('localhost','root','','wt_database');
-if (!$con)
-{
-    die('Could not connect: ' . mysqli_error($con));
-}
 $sql="SELECT * FROM clinic order by City,Town,CID ASC";
-$result = mysqli_query($con,$sql);
+$result = mysqli_query($conn,$sql);
 echo "<br><h2>TOTAL CLINICS IN DATABASE=<b>".mysqli_num_rows($result)."</b></h2><br>";
 echo "<table>
 <tr>
@@ -106,7 +102,7 @@ while($row = mysqli_fetch_array($result)) {
     echo "</tr>";
 }
 echo "</table>";
-mysqli_close($con);
+mysqli_close($conn);
 ?>
 </body>
 </html>
